perf_report/pyproject.toml
[tool.poetry]
name = "perf_report"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
tabulate = "^0.8.9"
click = "^8.0.4"
python-dateutil = "^2.8.2"
httpx = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
